+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 ## [1.0.0] - 2026-06-29
 * Base release
 
-## [1.0.1] - Unreleased
+## [1.1.0] - 2026-06-30
 * Added changelog
+* Added generation of T-free regions from a sequence, as well as visualization and summary stats
+* Added read length filtration in aggregate mode
 * Improved visualization of sequence Data Frame
-* Added backend support for generating T-free regions
+* Searching for a read by name is now easier
 * A read is now only searched for when the submit button is pressed, instead of any time the input changed